+++ src/mode-menu.js
@@ -8,7 +8,7 @@ let toggleKey = "I";
 
 function addModeMenu() {
   const container = document.querySelector(
-    '[class^="Editor_editor"] > .item__header > .header__extra-info > .hstack',
+    '[class^="Editor-module"] > .item__header > .header__extra-info > .hstack',
   );
 
   if (null === container) {